perf(charts): lazy-load chart-error diagnostics dialog (react-markdown)
